feat: refactor Employee class and adds EmployeeManager tests
