+++ lesson_15/tdd/tdd_app/src/main/java/com/codedifferently/lesson15/Employee.java
@@ -17,13 +17,6 @@ public class Employee {
     return id + ", " + name + ", " + department + "," + salary;
   }
 
-  public void setDetails(int id, String name, String department, double salary) {
-    this.id = id;
-    this.name = name;
-    this.department = department;
-    this.salary = salary;
-  }
-
   // Getters and setters
   public int getId() {
     return id;
